requirements.txt
streamlit
google-generativeai==0.5.4
requests
beautifulsoup4